\name{ssClust-package}
\alias{ssClust-package}
\docType{package}
\title{GMM for semi-supervised case a-la Mclust}
\description{Performs GMM for semi-supervised case a-la Mclust}
\details{
\tabular{ll}{
Package: \tab ssClust\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-20\cr
License: GPL (>=2)}

call to ssClust
}
\author{
Jordan Yoder

Jordan Yoder <jyoder6@jhu.edu>

}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{

}
\examples{
#see ssClust documentation
}
